+++ context/app/api/vitessce_confs/__init__.py
@@ -1,0 +1,1 @@
+from .assay_confs import get_view_config_class_for_data_types
